requirements.txt
# 深度学习框架
torch>=1.9.0

# 图处理 - PyTorch Geometric相关 (Windows兼容配置)
# 注意：在Windows上需要从pytorch-geometric官方wheel安装
# 安装命令: pip install torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.12.0+cpu.html
torch-geometric>=2.0.0
# torch-cluster>=1.5.0
# torch-scatter>=2.0.0  
# torch-sparse>=0.6.0

# 为了Windows兼容性，在代码中条件导入这些包

# 数据处理
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
h5py>=3.3.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 配置和日志
pyyaml>=5.4.0
omegaconf>=2.1.0
tensorboard>=2.7.0

# 工具包
tqdm>=4.62.0
pathlib2>=2.3.0

# 日志记录
loguru>=0.5.0

# 命令行参数解析
click>=8.0.0
argparse>=1.4.0

# 开发和测试工具
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-timeout>=1.4.0
pytest-xdist>=2.3.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# API服务
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
gunicorn>=20.1.0

# 数据验证
pydantic>=1.8.0
marshmallow>=3.12.0
schema>=0.7.0

# 性能监控
psutil>=5.8.0
memory-profiler>=0.58.0
line-profiler>=3.3.0

# 配置文件验证
cerberus>=1.3.0
jsonschema>=3.2.0

# 环境和部署
python-dotenv>=0.19.0
docker>=5.0.0
kubernetes>=18.0.0

# Jupyter支持
jupyter>=1.0.0
ipywidgets>=7.6.0
notebook>=6.4.0

# 医学影像处理 (可选)
SimpleITK>=2.1.0
nibabel>=4.0.0 